requirements.txt
alembic==1.7.4
amqp==5.1.1
bcrypt==3.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
blinker==1.4
cachelib==0.4.1
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
dnspython==1.16.0
dominate==2.6.0
email-validator==1.1.3
Flask==2.0.1
Flask-And-Redis==1.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
flask-redis==0.4.0
Flask-Reuploaded==1.2.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WhooshAlchemy==0.56
Flask-WTF==0.15.1
fpdf2==2.4.6
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.3.0
itsdangerous==2.0.1
Jinja2==3.0.1
kombu==5.2.4
lxml==4.6.3
Mako==1.1.5
MarkupSafe==2.0.1
mongoengine==0.23.1
packaging==21.3
pdfkit==1.0.0
Pillow==9.0.0
prompt-toolkit==3.0.29
pycparser==2.20
pymongo==3.12.0
pyparsing==3.0.6
python-dotenv==0.19.2
pytz==2022.1
redis==4.1.0
requests==2.26.0
rq==1.10.1
schedule==1.1.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.25
urllib3==1.26.7
vine==5.0.0
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==2.0.1
Whoosh==2.7.4
wrapt==1.13.3
WTForms==2.3.3
zipp==3.6.0
